+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Adds ability to filter column/table using a dropdown.
 - Adds feature that allows ember-tabular to wait until controller is ready for request to fire.
 - Add additional configurable wrapper classes for additional styling.
+- Add subcomponent to change the table limit/count.
 
 ### Changed
 - Update legacy name references in README.md.
@@ -22,3 +23,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Cleaned up code syntax to match airbnb's javascript style guide and eslint.
 - Fix visual filter layout bug to use component element wrappers and refactor component templates to reduce bloat.
 - When clearing filter of relationships, dot notation was not maintained.
+- Fix infinite request loop by ensuring that all ember-tabular shared properties are defined within controller. Added tesing coverage to verify this.